+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * refactor(`Context`): Returns null instead of throwing exception if the
   member is not found.
 * refactor(`Context`): Context is now `Traversable`.
+* feat(`MapperOptions`): Add context object to provide mapping options.
 
 ## 0.7.3